Update and rename src/net/minecraft/server/class_ael.java to src/net/minecraft/GameRules.java
